requirements.txt
uvicorn==0.32.1
fastapi==0.115.5
requests==2.32.3
pyotp==2.9.0
cryptography==43.0.3
python-dotenv==1.1.0
